[Spec] Factor out parsing logic
This is a small cleanup that makes it a bit easier to follow the logic
for finding a Range from a text directive by splitting out the logic
that parses the string into its individual components.

This is a first step in fixing and improving some of the issues in the
tree walking algorithm as part of #73.