[Spec] Explicitly mention 'focus' based attack
Another attack vector that's been brought to our attention is using the
fact that fragments (text or otherwise) cause focus to be applied. An
attacker can listen for notifications about losing focus in its own
document and infer that a fragment was successfully matched in a
cross-origin victim.

Our existing security measures mitigate this already, but this commit
adds some text to the non-normative motivation section to make readers
explicitly aware of this technique.