Relax word boundary requirement on prefix&suffix (#148)
* Relax word boundary requirement on prefix&suffix

As discussed in issue #137, there is no need to require the prefix to
end at a word boundary, or the suffix to start at a boundary.

* Simplify word boundary determination somewhat.

The existing definition is rather confusing, talking about code points
while using positions measured in code units; e.g. the “|endPosition|th
code point” actually refers to the code point that the |endPosition|th
code *unit* is part of.

This change avoids subtracting 1 from the substring length (which could
end up in the middle of a code point), and instead allows passing a
position that could equal the length of the string (which might strictly
speaking not qualify as a position in that string?). Instead of talking
only about word boundaries in relation to code points, I take the liberty
to talk about “the position of a boundary point”, which is analogous to
the position in a range (position N is right before code unit N, or if N
equals the string’s length it is after the last code unit).

I suppose this could be simplified further still if it is acceptable to
just test “if position is at a word boundary”, instead of having to first
find the first word boundary before/after the position.

* Added some non-normative text

* Review changes to previous commit

* Extend definition of word boundary position

* Create single word boundary test (for both start&end)

Less words, but it should do exactly the same.

* Tweak “is at word boundary” definition

Co-authored-by: David Bokan <bokan@chromium.org>