Update the algorithm for finding rangeEnd (#176)
This change fixes an issue in the find-range-from-text-directive algorithm of the spec. When searching for a rangeEnd with a suffix, a failing suffix match should continue searching the rest of the document for another instance of rangeEnd, rather than continuing the algorithm from the beginning.

See https://crbug.com/1171566 for the implementation fix.