Fix advancement in 'Next non-whitespace position'. (#192)
These steps didn't consider the case where they're run on a range whose
start is already at the end of its node.

Moving the node advancement step to the start of the loop fixes this.

Also consolidate the node advancement for various cases into a single
check. Previously the non-searchable subtree case would advance by 'next
node that isn’t a shadow-including descendant of node' which is now
omitted in the generalized case. However, this is equivalent since if
the next node is a shadow-including descendant of node, it must also be
in a non-searchable subtree and so it will keep advancing until it is no
longer a descendant of node.

Fixes #189