[Spec] Improve text in Navigating to a Text Fragment (#204)
The primary change is to make the document's indicated part return a range, both for text fragments and element fragments. An element fragment simply returns a range that spans the entire element.

Also change focus to be applied to the first node in the range, rather than the first common ancestor. This is important since the common ancestor could have sequential focus items that come before the range.