Improve integration with navigation (#225)
The fragment directive is now stripped from the URL whenever it is set to a history entry and the directive is kept separately in the entry's `directive state`.

The directive state may be shared between history entries where only the non-directive portion of the fragment changes.

Moving between session history entries will cause the directive state, if it differs from the current entry, to be set into then `uninvoked directives` member on the Document. The "scroll to the fragment" steps will read directives from this member and attempt to perform the text search.

When the fragment search ends, `uninvoked directives` is always cleared.

The changes here are primarily in the "3.3 The Fragment Directive" and "3.4 Text Directives" sections. Most changes in other sections are minor and/or text that was moved out into these two section (and heavily rewritten).

---------

Co-authored-by: Jake Archibald <jaffathecake@gmail.com>