Fix application of force-load-at-top
Currently force-load-at-top is checked in the 'scroll to the fragment'
steps. This is wrong since it shouldn't apply on same-document
navigations and it should also affect history scroll restoration.

Move this check to happen in the 'update document for history step
application' where history scroll is restored and where we can
differentiate between new- and same-document cases.

Fixes #186