Change the search algorithm as described in #262.
This patch updates the `find a string in range` algorithm
to use a new boolean parameter `matchMustBeAtBeginning`,
which will early-return the algorithm if the `query` is not
at the beginning of the `searchRange`.
This variant is used when searching for the `start` parameter
when a `prefix` is present, and for the `suffix` parameter.

This also required some changes to the control flow of the
`find a range from a text directive` algorithm.